Fix for loading Fieldtrip
Similar to the fix for EEGLAB files, this is an update for importing fieldtrip files - we had issues with reading in fieldtrip files created using matlab Version 7.0 (-v7.3). There were some format differences (perhaps due to a matlab version update?) - that caused issues in reading some metadata using h5py..
This also includes update to the fieldtrip.py for padding the dataset so that the GUI doesn't crash when reaching the end of the recording. These changes seem to resolve the issue.