(Duplicate) Minor changes to prevent crashes in export event params
Hey hey!

I am running into issues with the same  part of export_event_params() as a couple of months ago.

I made a double check over this, and this edit seems to work over as many scenarios as I could think of.

ALSO: could this please be released through pypi as a new version? Thanks!